requirements.txt
amqp==5.3.1
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.3
black==25.12.0
cachetools==6.2.2
cbor2==5.7.1
celery==5.5.3
certifi==2025.11.12
certvalidator==0.11.1
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
decorator==5.2.1
defusedxml==0.7.1
Django==6.0
django-appconf==1.2.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-imagekit==6.0.0
django-polymorphic==4.1.0
django-sslserver==0.22
django-webauthn==0.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
google-auth==2.45.0
hyperlink==21.0.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Incremental==24.11.0
isort==7.0.0
kombu==5.5.4
Markdown==3.10
mccabe==0.7.0
moviepy==2.2.1
msgpack==1.1.2
mypy_extensions==1.1.0
numpy==2.3.5
oscrypto==1.3.0
packaging==25.0
pathspec==0.12.1
pilkit==3.0
pillow==11.3.0
pillow_heif==1.1.1
platformdirs==4.5.1
proglog==0.1.12
prompt_toolkit==3.0.52
py-ubjson==0.16.1
py-webauthn==0.0.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
redis==7.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webauthn==2.7.0
wheel==0.45.1
zope.interface==8.1.1
